Add volume multi attach support
This patch includes the Cinder changes needed
to support volume multiple attaches.  Nova and
python-cinderclient also need patches associated
to provide support for multiple attachments.

This adds the multiattach flag to volumes.  When a
volume is created, a multiattach flag can be set,
which allows a volume to be attached to more than
one Nova instance or host.  If the multiattach flag is
not set on a volume, it cannot be attached to more
than one Nova instance or host

Each volume attachment is tracked in a
new volume_attachment table.  The attachment id is
the unique identifier for each attachment to an
instance or host.

When a volume is to be detached the attachment
uuid must be passed in to the detach call in
order to determine which attachment should be
removed.  Since a volume can be attached to an
instance and a host, the attachment id is used
as the attachment identifier.

Nova:
https://review.openstack.org/#/c/153033/
https://review.openstack.org/#/c/153038/

python-cinderclient:
https://review.openstack.org/#/c/85856/

Change-Id: I950fa00ed5a30e7758245d5b0557f6df42dc58a3
Implements: blueprint multi-attach-volume
APIImpact

(cherry picked from commit 10d54216871fde27172fbcb1a3c9bfec59b824c2)

Conflicts:
	cinder/api/v1/volumes.py
	cinder/backup/manager.py
	cinder/db/sqlalchemy/api.py
	cinder/objects/volume.py
	cinder/tests/api/v2/test_volumes.py
	cinder/tests/test_hp3par.py
	cinder/tests/test_migrations.py
	cinder/tests/test_volume.py
	cinder/volume/api.py
	cinder/volume/drivers/san/hp/hp_3par_common.py
	cinder/volume/drivers/san/hp/hp_3par_fc.py
	cinder/volume/drivers/san/hp/hp_3par_iscsi.py
	cinder/volume/flows/api/create_volume.py
	cinder/volume/manager.py
	cinder/volume/rpcapi.py