Move retype quota checks to API
Currently the retype quota check is done in the manager, where it will
raise an exception if the check fails. This patch moves the check to the
API so that the user is given feedback before the RPC cast is made.

APIImpact
Closes-bug: #1508249

Change-Id: Iba24edd8ad824837028353b52c90742df55c9173
(cherry picked from commit d78399e614a90b062bcc1ff10aa00a037b158cad)