restore volume status in create_backup when backup status is not correct.
When creating backup, if backup status is not correct, an exception of
InvalidBackup will be raised, and volume status will be set to
'available'. This behavior is good before we added non disruptive
backup feature. But when doing a force backup, volume status should be
restored to 'in-use'. So the right way is just restoring volume status
to the attribute of 'previous_status'.

Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>