take care of non-disruptive backup when detaching a volume.
Volume actions like 'os-detach' and 'os-force_detach' will in the end
call db.volume_detached, which update volume status to 'available'
directly. If at the same time, a backing-up is on the way, when backup
is finished, volume status will be restored to 'in-use' other than
'available'.

So we should not change volume status when detaching a volume if there's
an ongoing backing-up, only update previous_status. Then when backing-up
is finished, volume previous_status should be checked again.

Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>