only allow force create snapshot when volume is usable.
Volume deleting may take a while, but create_snapshot_force
ignors volume status. So if a snapshot is created while the
volume is deleting, when the deletion is finished, this newly
created snapshot will be left totally unusable. To resolve
this, just allow force create snapshot if volume status is
'available' or 'in-use' just like force backup.

Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>