Correctly open rbd image in ceph backup driver.
Currently in _check_restore_vol_size of ceph backup driver, destination
rbd volume is opened by RADOSClient with authentication of cinder backup
user. This will lead to PermissionError.

In fact destination rbd volume is already opened as a RBDImageIOWrapper
object, which is also passed to cinder-backup restore function, so we
should just use it.

Change-Id: Ibdfd0e0b09410eb47e3ebc6ce61d4473cf0d5677
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>