Correctly reset volume status while cinder-backup restarts.
Before this patch, when cinder-backup restarts, volumes and backups are
resetting status separately. But volumes are filtered by host,
cinder-backup's host configuration is not the same as cinder-volume by
default, so in fact no volumes will be found and reset.

This patch comes from faee0520a455e3402dd7d77662b077b64956db08, volumes
are checked and reset according to the incompleted backup.

Fixes: http://192.168.15.2/issues/9978

Change-Id: I045c7711039ce4fa656c9fde01e6eb5329fbafcd
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>